+++ at_client/lib/src/client/remote_secondary.dart
@@ -3,11 +3,8 @@ import 'dart:io';
 
 import 'package:at_client/at_client.dart';
 import 'package:at_client/src/client/secondary.dart';
-import 'package:at_client/src/preference/at_client_preference.dart';
 import 'package:at_commons/at_builders.dart';
 import 'package:at_lookup/at_lookup.dart';
-import 'package:at_lookup/src/connection/outbound_connection.dart';
-import 'package:at_utils/at_logger.dart';
 import 'package:at_utils/at_utils.dart';
 import 'package:internet_connection_checker/internet_connection_checker.dart';
 
@@ -17,7 +14,7 @@ class RemoteSecondary implements Secondary {
 
   late String _atSign;
 
-  late var _preference;
+  late AtClientPreference _preference;
 
   late AtLookupImpl atLookUp;
 
@@ -34,7 +31,7 @@ class RemoteSecondary implements Secondary {
   /// Optionally [privateKey] is passed for verb builders which require authentication.
   @override
   Future<String> executeVerb(VerbBuilder builder, {sync = false}) async {
-    var verbResult;
+    String verbResult;
     verbResult = await atLookUp.executeVerb(builder);
     return verbResult;
   }
@@ -46,7 +43,7 @@ class RemoteSecondary implements Secondary {
   }
 
   Future<String?> executeCommand(String atCommand, {bool auth = false}) async {
-    var verbResult;
+    String? verbResult;
     verbResult = await atLookUp.executeCommand(atCommand, auth: auth);
     return verbResult;
   }
@@ -64,7 +61,7 @@ class RemoteSecondary implements Secondary {
 
   /// Generates digest using from verb response and [secret] and performs a CRAM authentication to
   /// secondary server
-  Future<bool> authenticate_cram(var secret) async {
+  Future<bool> authenticateCram(var secret) async {
     var authResult = await atLookUp.authenticate_cram(secret);
     return authResult;
   }
@@ -98,7 +95,7 @@ class RemoteSecondary implements Secondary {
       var host = secondaryInfo[0];
       var port = secondaryInfo[1];
       var internetAddress = await InternetAddress.lookup(host);
-      //#TODO getting first ip for now. explore best solution
+      //TODO getting first ip for now. explore best solution
       var addressCheckOptions =
           AddressCheckOptions(internetAddress[0], port: int.parse(port));
       return (await InternetConnectionChecker()