+++ at_client/lib/src/response/notification_response_parser.dart
@@ -1,31 +1,26 @@
-import 'dart:convert';
-
 import 'package:at_client/src/response/at_notification.dart';
 import 'package:at_client/src/response/default_response_parser.dart';
+import 'package:at_client/src/response/json_utils.dart';
 import 'package:at_client/src/response/response.dart';
 
 class NotificationResponseParser extends DefaultResponseParser {
-  @override
-  Response parse(String responseString) {
-    return super.parse(responseString);
-  }
-
-  List<AtNotification> getAtNotifications(Response response) {
+  List<AtNotification> getAtNotifications(AtResponse response) {
     final notificationList = <AtNotification>[];
     if (response.isError) {
       return [];
     }
     final notificationJson = response.response;
     var notifications = notificationJson.split('notification: ');
-    notifications.forEach((notification) {
+    for (var notification in notifications) {
       if (notification.isEmpty) {
-        return;
+        continue;
       }
       notification = notification.replaceFirst('notification:', '');
       notification = notification.trim();
-      final atNotification = AtNotification.fromJson(jsonDecode(notification));
+      final atNotification =
+          AtNotification.fromJson(JsonUtils.decodeJson(notification));
       notificationList.add(atNotification);
-    });
+    }
     return notificationList;
   }
 }