+++ at_client/lib/src/service/file_transfer_service.dart
@@ -1,9 +1,9 @@
 import 'dart:io';
 
+import 'package:archive/archive_io.dart';
 import 'package:at_client/src/stream/file_transfer_object.dart';
 import 'package:at_client/src/util/constants.dart';
 import 'package:http/http.dart' as http;
-import 'package:archive/archive_io.dart';
 import 'package:http/http.dart';
 
 class FileTransferService {
@@ -11,7 +11,7 @@ class FileTransferService {
       List<int> file, String container, String fileName) async {
     try {
       var response = await http.post(
-        Uri.parse(TextConstants.FILEBIN_URL + '$container/' + fileName),
+        Uri.parse(TextConstants.fileBinURL + '$container/' + fileName),
         body: file,
       );
       return response;
@@ -24,7 +24,7 @@ class FileTransferService {
       File file, String container, String fileName) async {
     try {
       var postUri =
-          Uri.parse(TextConstants.FILEBIN_URL + '$container/' + fileName);
+          Uri.parse(TextConstants.fileBinURL + '$container/' + fileName);
       final streamedRequest = http.StreamedRequest('POST', postUri);
 
       streamedRequest.contentLength = await file.length();