+++ at_client/lib/src/service/notification_service_impl.dart
@@ -4,11 +4,9 @@ import 'dart:convert';
 import 'package:at_client/at_client.dart';
 import 'package:at_client/src/exception/at_client_exception_util.dart';
 import 'package:at_client/src/listener/at_sign_change_listener.dart';
-import 'package:at_client/src/listener/connectivity_listener.dart';
 import 'package:at_client/src/listener/switch_at_sign_event.dart';
 import 'package:at_client/src/manager/monitor.dart';
 import 'package:at_client/src/preference/monitor_preference.dart';
-import 'package:at_client/src/response/at_notification.dart';
 import 'package:at_client/src/response/notification_response_parser.dart';
 import 'package:at_client/src/service/notification_service.dart';
 import 'package:at_commons/at_commons.dart';
@@ -17,7 +15,7 @@ import 'package:at_utils/at_logger.dart';
 class NotificationServiceImpl
     implements NotificationService, AtSignChangeListener {
   Map<String, StreamController> streamListeners = {};
-  final EMPTY_REGEX = '';
+  final emptyRegex = '';
   static const notificationIdKey = '_latestNotificationId';
   static final Map<String, NotificationService> _notificationServiceMap = {};
 
@@ -26,7 +24,7 @@ class NotificationServiceImpl
   late AtClient _atClient;
   Monitor? _monitor;
   ConnectivityListener? _connectivityListener;
-  var _lastMonitorRetried;
+  dynamic _lastMonitorRetried;
 
   static Future<NotificationService> create(AtClient atClient) async {
     if (_notificationServiceMap.containsKey(atClient.getCurrentAtSign())) {
@@ -59,7 +57,7 @@ class NotificationServiceImpl
   }
 
   Future<void> _startMonitor() async {
-    if (_monitor != null && _monitor!.status == MonitorStatus.Started) {
+    if (_monitor != null && _monitor!.status == MonitorStatus.started) {
       _logger.finer(
           'monitor is already started for ${_atClient.getCurrentAtSign()}');
       return;
@@ -73,16 +71,19 @@ class NotificationServiceImpl
         MonitorPreference()..keepAlive = true,
         _monitorRetry);
     await _monitor!.start(lastNotificationTime: lastNotificationTime);
-    if (_monitor!.status == MonitorStatus.Started) {
+    if (_monitor!.status == MonitorStatus.started) {
       _isMonitorPaused = false;
     }
   }
 
   Future<int?> _getLastNotificationTime() async {
-    final atValue = await _atClient.get(AtKey()..key = notificationIdKey);
-    if (atValue.value != null) {
-      _logger.finer('json from hive: ${atValue.value}');
-      return jsonDecode(atValue.value)['epochMillis'];
+    var atKey = AtKey()..key = notificationIdKey;
+    if (_atClient.getLocalSecondary()!.keyStore!.isKeyExists(atKey.key!)) {
+      final atValue = await _atClient.get(atKey);
+      if (atValue.value != null) {
+        _logger.finer('json from hive: ${atValue.value}');
+        return jsonDecode(atValue.value)['epochMillis'];
+      }
     }
     return null;
   }
@@ -103,14 +104,14 @@ class NotificationServiceImpl
       final notificationParser = NotificationResponseParser();
       final atNotifications = notificationParser
           .getAtNotifications(notificationParser.parse(notificationJSON));
-      atNotifications.forEach((atNotification) async {
+      for (var atNotification in atNotifications) {
         // Saves latest notification id to the keys if its not a stats notification.
         if (atNotification.id != '-1') {
           await _atClient.put(AtKey()..key = notificationIdKey,
               jsonEncode(atNotification.toJson()));
         }
         streamListeners.forEach((regex, streamController) {
-          if (regex != EMPTY_REGEX) {
+          if (regex != emptyRegex) {
             if (regex.allMatches(atNotification.key).isNotEmpty) {
               streamController.add(atNotification);
             }
@@ -118,7 +119,7 @@ class NotificationServiceImpl
             streamController.add(atNotification);
           }
         });
-      });
+      }
     } on Exception catch (e) {
       _logger.severe(
           'exception processing: error:${e.toString()} notificationJson: $notificationJSON');
@@ -152,7 +153,7 @@ class NotificationServiceImpl
       {Function? onSuccess, Function? onError}) async {
     var notificationResult = NotificationResult()
       ..atKey = notificationParams.atKey;
-    var notificationId;
+    dynamic notificationId;
     try {
       // Notifies key to another notificationParams.atKey.sharedWith atsign
       // Returns the notificationId.
@@ -204,9 +205,9 @@ class NotificationServiceImpl
   /// Queries the status of the notification
   /// Takes the notificationId as input as returns the status of the notification
   Future<String> _getFinalNotificationStatus(String notificationId) async {
-    var status;
+    String status = '';
     // For every 2 seconds, queries the status of the notification
-    while (status == null || status == 'data:queued') {
+    while (status.isEmpty || status == 'data:queued') {
       await Future.delayed(Duration(seconds: 2),
           () async => status = await _atClient.notifyStatus(notificationId));
     }
@@ -215,7 +216,7 @@ class NotificationServiceImpl
 
   @override
   Stream<AtNotification> subscribe({String? regex}) {
-    regex ??= EMPTY_REGEX;
+    regex ??= emptyRegex;
     if (streamListeners.containsKey(regex)) {
       _logger.finer('subscription already exists');
       return streamListeners[regex]!.stream as Stream<AtNotification>;