+++ at_client/lib/src/util/constants.dart
@@ -1,4 +1,4 @@
 class TextConstants {
-  static const String FILEBIN_URL = 'https://ck6agzxiog6kmb.atsign.com/';
-  static const String FILE_TRANSFER_KEY = 'file_transfer_';
+  static const String fileBinURL = 'https://ck6agzxiog6kmb.atsign.com/';
+  static const String fileTransferKey = 'file_transfer_';
 }