+++ at_client_mobile/lib/src/at_client_service.dart
@@ -1,12 +1,8 @@
-import 'dart:async';
 import 'dart:convert';
 import 'dart:core';
 
-import 'package:at_client/at_client.dart';
 import 'package:at_client_mobile/at_client_mobile.dart';
-import 'package:at_client_mobile/src/at_client_auth.dart';
 import 'package:at_client_mobile/src/auth_constants.dart';
-import 'package:at_client_mobile/src/onboarding_status.dart';
 import 'package:at_commons/at_builders.dart';
 import 'package:at_commons/at_commons.dart';
 import 'package:at_lookup/at_lookup.dart';
@@ -345,8 +341,7 @@ class KeychainUtil {
   }
 
   static Future<String?> getAESKey(String atsign) async {
-    return await _keyChainManager.getValue(
-        atsign, keychainSelfEncryptionKey);
+    return await _keyChainManager.getValue(atsign, keychainSelfEncryptionKey);
   }
 
   static Future<String?> getSelfEncryptionKey(String atSign) async {